Put Shop in separate file
